Reorganize Bowser Jr Statuses
